// frontend/src/lib/generated/models/BaseImageRequest.ts
export interface BaseImageRequest {
    /**
     * 
     * @type {number}
     * @memberof BaseImageRequest
     */
    baseImageId?: number;
    /**
     * 
     * @type {BaseImageRequestBaseImage}
     * @memberof BaseImageRequest
     */
    baseImage?: BaseImageRequestBaseImage;
    /**
     * 
     * @type {BaseImageType}
     * @memberof BaseImageRequest
     */
    type: BaseImageType;
    /**
     * 
     * @type {string}
     * @memberof BaseImageRequest
     */
    value: string;
    /**
     * 
     * @type {BaseImageRequestGit}
     * @memberof BaseImageRequest
     */
    git?: BaseImageRequestGit;
    /**
     * 
     * @type {BaseImageRequestExe}
     * @memberof BaseImageRequest
     */
    exe?: BaseImageRequestExe;
    /**
     * 
     * @type {Array<BaseImageArgument>}
     * @memberof BaseImageRequest
     */
    arguments: Array<BaseImageArgument>;
}