// frontend/src/lib/generated/models/BaseImageType.ts
/* tslint:disable */
/* eslint-disable */
/**
 * seaport.com
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const BaseImageType = {
    Git: 'GIT',
    Exe: 'EXE'
} as const;
export type BaseImageType = typeof BaseImageType[keyof typeof BaseImageType];


export function instanceOfBaseImageType(value: any): boolean {
    for (const key in BaseImageType) {
        if (Object.prototype.hasOwnProperty.call(BaseImageType, key)) {
            if (BaseImageType[key as keyof typeof BaseImageType] === value) {
                return true;
            }
        }
    }
    return false;
}

export function BaseImageTypeFromJSON(json: any): BaseImageType {
    return BaseImageTypeFromJSONTyped(json, false);
}

export function BaseImageTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): BaseImageType {
    return json as BaseImageType;
}

export function BaseImageTypeToJSON(value?: BaseImageType | null): any {
    return value as any;
}

export function BaseImageTypeToJSONTyped(value: any, ignoreDiscriminator: boolean): BaseImageType {
    return value as BaseImageType;
}